Move EphemeralStore to unit tests
